+++ raiden-dapp/src/model/types.ts
@@ -8,6 +8,10 @@ export interface Token {
   readonly name?: string;
 }
 
+export interface Presences {
+  [address: string]: boolean;
+}
+
 export interface AccTokenModel {
   name: string;
   symbol: string;