+++ raiden-dapp/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- [#218] Matrix availability check to address inputs
+
 ### Changed
 - User can now dismiss/hide the transfer progress dialog.
 - Fix utility token display on PFS route selection.